加入aop支持、加入日志，但并没有完全搞懂，至少能用了

+++ src/main/java/com/stkj/allpay/menu/repository/MenuRespository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.stkj.allpay.menu.entity.Menu;
 
-public interface MenuRespository extends JpaRepository<Menu, Long>{
+public interface MenuRespository extends JpaRepository<Menu, Integer>{
 
-//	List<Menu> findAll();
+	
 }
